Add reference field to capture
In a multi-capture scenario there are more than one capture request for the same orderCode. When the asynchronous order notification comes back from Worldpay the is no way to tell which capture initiated the notification.

This change solves that by adding a reference field to the capture request which will be sent back in the WP notification.